+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Warm starting for collision solving (#1248, **@fallenatlas**).
 - Conditional shader compilation (#1406, **@tomas7770**).
 - Collision layers and masks, for filtering collisions and raycasts (**@RiscadoA**).
+- Frustum geometry and intersection functions (#1184, **@mkuritsu**).
 
 ### Changed